headerIconv
- Make the spec as follows
...
Summary: default summary
Summary(ja_JP.eucJP): [some strings encoded in euc-jp]
Name:hogefuga
...
- Show summary in each locale.
$ LANG=ja_JP.utf8 rpm -qp --qf "%{SUMMARY}\n" hogefuga.rpm
[some strings encoded in euc-jp]

do not display in UTF-8.

This patch improve the parser of spec-file which are written in multi-languages.
In case the language code is ja_JP , as you know ja_JP use some different encodings
such as ja_JP.eucJP,ja_JP.sjis,ja_JP.utf8...
When the spec-file are written in eucJP but the local LANG environment is UTF-8,
rpm will display encoded characters not same as LANG environment(a.k.a Mojibake).
So this patch will convert the data to a specific encoding which used in the selected locale.