speedup by not checking same files with different paths through symlink
without this patch, "rpm -e" or "rpm -U" will need to stat(2) every dirnames
of files from the package (eg COPYING) in the db. It will also need to access
all the packages in "Packages" db that have the same basename.
This is quite costly when not in cache
(eg on a test here: >300 stats, and so 3 seconds after a "echo 3 > /proc/sys/vm/drop_caches")
this breaks urpmi test case test_rpm_i_fail('gd') in superuser--file-conflicts.t,
but this is bad design anyway

- Dirnames db can't be used together with Basenames to construct the full
filename since dirindexes are only available in Packages db. So Dirnames db is
only used to know if the directory exists in the package. It's quite effective
anyway since for cases like basename "COPYING" or "README", the dirname is
specific to one package.

- rpmdbGet_dbiIndexSet() is created out of rpmdbGrowIterator()

- dbiIntersectSet() is mostly a copy/paste of dbiPruneSet()

- set_mi_string_key() is a small helper function to keep code clean